# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BookingCreateRequest": {
        "properties": {
          "fulfillment_method": {
            "enum": [
              "pickup",
              "owner_delivery"
            ],
            "title": "Fulfillment Method",
            "type": "string"
          },
          "quote_id": {
            "format": "uuid",
            "title": "Quote Id",
            "type": "string"
          }
        },
        "required": [
          "quote_id",
          "fulfillment_method"
        ],
        "title": "BookingCreateRequest",
        "type": "object"
      },
      "BookingResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "deposit_minor": {
            "title": "Deposit Minor",
            "type": "integer"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "fulfillment_method": {
            "title": "Fulfillment Method",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "listing_id": {
            "format": "uuid",
            "title": "Listing Id",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "public_number": {
            "title": "Public Number",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "rental_charge_minor": {
            "title": "Rental Charge Minor",
            "type": "integer"
          },
          "renter_user_id": {
            "format": "uuid",
            "title": "Renter User Id",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "public_number",
          "listing_id",
          "renter_user_id",
          "owner_user_id",
          "starts_at",
          "ends_at",
          "quantity",
          "status",
          "fulfillment_method",
          "rental_charge_minor",
          "deposit_minor",
          "currency",
          "created_at"
        ],
        "title": "BookingResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageCreateRequest": {
        "properties": {
          "body": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "client_message_id": {
            "format": "uuid",
            "title": "Client Message Id",
            "type": "string"
          }
        },
        "required": [
          "client_message_id",
          "body"
        ],
        "title": "MessageCreateRequest",
        "type": "object"
      },
      "MessageResponse": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sender_user_id": {
            "format": "uuid",
            "title": "Sender User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sender_user_id",
          "body",
          "created_at"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "QuoteRequest": {
        "properties": {
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "quantity": {
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Quantity",
            "type": "integer"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "unit": {
            "enum": [
              "hour",
              "day",
              "week",
              "month"
            ],
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "starts_at",
          "ends_at",
          "quantity",
          "unit"
        ],
        "title": "QuoteRequest",
        "type": "object"
      },
      "QuoteResponse": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "deposit_minor": {
            "title": "Deposit Minor",
            "type": "integer"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "listing_id": {
            "format": "uuid",
            "title": "Listing Id",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "rental_charge_minor": {
            "title": "Rental Charge Minor",
            "type": "integer"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "listing_id",
          "starts_at",
          "ends_at",
          "quantity",
          "rental_charge_minor",
          "deposit_minor",
          "currency",
          "expires_at"
        ],
        "title": "QuoteResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "4by4 For Hire API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bookings": {
      "get": {
        "operationId": "list_bookings_api_v1_bookings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingResponse"
                  },
                  "title": "Response List Bookings Api V1 Bookings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Bookings",
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "create_booking_api_v1_bookings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}": {
      "get": {
        "operationId": "get_booking_api_v1_bookings__booking_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/accept": {
      "post": {
        "operationId": "accept_booking_api_v1_bookings__booking_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/cancel": {
      "post": {
        "operationId": "cancel_booking_api_v1_bookings__booking_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Booking",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/api/v1/bookings/{booking_id}/messages": {
      "get": {
        "operationId": "list_messages_api_v1_bookings__booking_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response List Messages Api V1 Bookings  Booking Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Messages",
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "send_message_api_v1_bookings__booking_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Message",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/api/v1/bookings/{booking_id}/reject": {
      "post": {
        "operationId": "reject_booking_api_v1_bookings__booking_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Booking",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/api/v1/listings/{listing_id}/quotes": {
      "post": {
        "operationId": "create_quote_api_v1_listings__listing_id__quotes_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Quote",
        "tags": [
          "bookings"
        ]
      }
    },
...
  }
}